backend/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
